fix(loader): upcast KDA A_log/dt_bias to fp32 for inference
KDA kernels require fp32 gating params; training feeds them via autocast,
so a bf16 inference model crashes on long sequences (flash_kda backend:
'A_log must be float32'). Short prompts take a different path and hide
the bug -- L1 ran 256 problems before hitting it, code few-shot (longer
demos) crashed on the first batch. Upcast after the dtype conversion so
it sticks; code_fewshot now passes dtype= to load_checkpoint instead of
a separate .to() that would undo it. L1/math_zh need the same switch on
KDA checkpoints.

+++ eval/code_fewshot.py
@@ -111,8 +111,7 @@ def main():
 
     import fone  # noqa: F401
 
-    model, cfg = load_checkpoint(args.ckpt, device=args.device)
-    model = model.to(torch.bfloat16)
+    model, cfg = load_checkpoint(args.ckpt, device=args.device, dtype=torch.bfloat16)
     tok = load_tokenizer(args.tokenizer, cfg)
     fone_on = getattr(cfg, "fone", False)
     num_id = getattr(cfg, "num_id", None)
